+++ hydromt/data_adapter/dataframe.py
@@ -54,10 +54,10 @@ class DataFrameAdapter(DataAdapter):
             Path to data source. If the dataset consists of multiple files, the path may
             contain {variable}, {year}, {month} placeholders as well as path
             search pattern using a '*' wildcard.
-        driver: {'csv', 'xlsx', 'xls', 'fwf'}, optional
+        driver: {'csv', 'parquet', 'xlsx', 'xls', 'fwf'}, optional
             Driver to read files with, for 'csv' :py:func:`~pandas.read_csv`,
-            for {'xlsx', 'xls'} :py:func:`~pandas.read_excel`, and for 'fwf'
-            :py:func:`~pandas.read_fwf`.
+            for 'parquet' :py:func:`~pandas.read_parquet`, for {'xlsx', 'xls'}
+            :py:func:`~pandas.read_excel`, and for 'fwf' :py:func:`~pandas.read_fwf`.
             By default the driver is inferred from the file extension and falls back to
             'csv' if unknown.
         filesystem: {'local', 'gcs', 's3'}, optional
@@ -131,7 +131,7 @@ class DataFrameAdapter(DataAdapter):
         data_name : str
             Name of the output file without extension.
         driver : str, optional
-            Driver to write the file, e.g., 'csv', 'excel'. If None,
+            Driver to write the file, e.g., 'csv','parquet', 'excel'. If None,
             the default behavior is used.
         variables : list of str, optional
             Names of DataFrame columns to include in the output. By default,
@@ -172,6 +172,9 @@ class DataFrameAdapter(DataAdapter):
             fn_out = join(data_root, f"{data_name}.csv")
             obj.to_csv(fn_out, **kwargs)
             kwargs["index_col"] = obj.index.name
+        elif driver == "parquet":
+            fn_out = join(data_root, f"{data_name}.parquet")
+            obj.to_parquet(fn_out, **kwargs)
         elif driver == "excel":
             fn_out = join(data_root, f"{data_name}.xlsx")
             obj.to_excel(fn_out, **kwargs)
@@ -211,6 +214,8 @@ class DataFrameAdapter(DataAdapter):
 
         if self.driver in ["csv"]:
             df = pd.read_csv(self.path, **kwargs)
+        elif self.driver == "parquet":
+            df = pd.read_parquet(self.path, **kwargs)
         elif self.driver in ["xls", "xlsx", "excel"]:
             df = pd.read_excel(self.path, engine="openpyxl", **kwargs)
         elif self.driver in ["fwf"]: